src/authservice/requirements.txt
grpcio==1.66.0
grpcio-tools==1.66.0
protobuf==5.27.2
bcrypt==4.1.3
psycopg2-binary==2.9.9
six==1.16.0
typing_extensions==4.12.2
pytest==8.3.2